NumPy-Assignment
NumPy is a Python library used for scientific computing and numerical operations. It provides powerful data structures for multi-dimensional arrays and matrices, as well as a large collection of mathematical functions to operate on these arrays.
features of NumPy include:
1) ndarray: an efficient and versatile n-dimensional array object for fast numerical operations on large datasets.
2) Broadcasting: a powerful mechanism that allows for arithmetic operations between arrays of different shapes and sizes.
3) Vectorization: the ability to perform element-wise operations on entire arrays without explicit loops.
4)Universal Functions (ufuncs): a collection of functions that operate element-wise on ndarrays and support broadcasting.
5) Linear algebra, Fourier transforms, and random number capabilities.

NumPy is a foundational library for many scientific computing applications in Python, and is often used in conjunction with other libraries such as SciPy, Pandas, and Matplotlib.

To use NumPy, you can install it using pip:

!pip install numpy         (Jupyter notebook)

Once installed, you can import it in your Python code with:

import numpy as np

From here, you can use NumPy's powerful array manipulation and mathematical functions in your code.