List data type program
This file contain simple program related to list data type and use of pickle 
Pickle is a Python module that allows you to serialize (convert to a byte stream) and deserialize (convert from a byte stream) Python objects. This module provides a convenient way to store Python objects in a file or transfer them over a network, and is commonly used for data serialization, configuration management, and other similar purposes.

Pickle can handle most Python objects, including built-in types (such as integers, floats, and lists) and user-defined classes and objects. When you pickle an object, the resulting byte stream contains all the information needed to reconstruct the object at a later time, including its internal state and any references to other objects.

The main advantage of using pickle is that it allows you to store complex Python data structures as a single object, which can be easily saved to disk or sent over a network. This is particularly useful when working with large datasets or complex objects, as it provides a simple and efficient way to manage and manipulate the data.

However, it is important to note that pickle is not a secure serialization format, as it can execute arbitrary code during deserialization. 
Therefore, you should only use pickle to serialize and deserialize data from trusted sources or use alternative serialization formats such as JSON or XML when working with untrusted data.